move methods for computing clusters' similarity
