+++ modules/chameleon/src/test/java/org/clueminer/chameleon/GraphClusterTest.java
@@ -16,6 +16,7 @@
  */
 package org.clueminer.chameleon;
 
+import org.clueminer.chameleon.similarity.RiRcSimilarity;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import org.clueminer.dataset.api.Dataset;
